syntax highligh started developing  for many file extensions. They msut be in a map

+++ src/main.cpp
@@ -1,5 +1,7 @@
 #include "Headers/mainwindow.h"
 #include <QApplication>
+#include <QErrorMessage>
+
 
 int main(int argc, char *argv[])
 {
@@ -7,11 +9,22 @@ int main(int argc, char *argv[])
     MainWindow w;
     w.show();
 
-    QFile File("css/dark.qss");
-    File.open(QFile::ReadOnly);
-    QString StyleSheet = QLatin1String(File.readAll());
 
-    a.setStyleSheet(StyleSheet);
+    QFile File("css/dark.qss");
+
+    if (File.exists()) {
+        try{
+            File.open(QFile::ReadOnly);
+            QString StyleSheet = QLatin1String(File.readAll());
+
+            a.setStyleSheet(StyleSheet);
+        }
+        catch (...) {
+            QErrorMessage *fileError = new QErrorMessage();
+            fileError->showMessage("Stylesheet file not found");
+        }
+    }
+
 
     return a.exec();
 }
